% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BV_fishline_2_rdbes_1_17.R
\name{BV_fishline_2_rdbes}
\alias{BV_fishline_2_rdbes}
\title{FishLine 2 RDBES, biological variables (BV)}
\usage{
BV_fishline_2_rdbes(
  data_model = data_model,
  years = c(2016),
  cruises = c("MON", "SEAS"),
  type = "individual",
  species_for_bv = c("TOR", "HVL")
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{cruises}{Name of cruises in national database}

\item{species_for_bv}{}
}
\value{

}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.17. This function ...
}
\examples{



}
