% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_fishline_2_rdbes_1_17.R
\name{DE_fishline_2_rdbes}
\alias{DE_fishline_2_rdbes}
\title{FishLine 2 RDBES, design (DE)}
\usage{
DE_fishline_2_rdbes(
  data_model = data_model,
  year = 2016,
  type = "observer_at_sea",
  hierarchy = 1,
  strata = "quarter"
)
}
\arguments{
\item{type}{observer_at_sea |}

\item{stata}{quarter | none}
}
\value{

}
\description{
FishLine 2 RDBES, design (DE)
}
