% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishline2rdbes-package.R
\docType{package}
\name{fishline2rdbes-package}
\alias{fishline2rdbes}
\alias{fishline2rdbes-package}
\title{fishline2rdbes: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
