% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FT_fishline_2_rdbes.R
\name{FT_fishline_2_rdbes}
\alias{FT_fishline_2_rdbes}
\title{FishLine 2 RDBES, Fishing trip (FT)}
\usage{
FT_fishline_2_rdbes(
  ref_path = "Q:/mynd/RDB/create_RDBES_data/references",
  encryptedVesselCode_path =
    "Q:/mynd/RDB/create_RDBES_data/RDBES_data_call_2021/output/for_production",
  years = 2016,
  sampling_scheme = "DNK_AtSea_Observer_Active",
  type = "everything"
)
}
\arguments{
\item{type}{only_mandatory | everything}

\item{path_to_data_model_baseTypes}{Where to find the baseTypes for the data model}

\item{year}{years needed}

\item{cruises}{Name of cruises in national database}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
