% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_model.R
\name{get_data_model}
\alias{get_data_model}
\title{Title}
\usage{
get_data_model(
  table_name,
  data_model_path =
    "Q:/20-forskning/20-dfad/data/Data/RDBES/sample_data/create_RDBES_data/input",
  data_model_version = "v1_19_60"
)
}
\arguments{
\item{data_model_version}{}
}
\description{
Title
}
