% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gear_info.R
\name{get_gear_info}
\alias{get_gear_info}
\title{Get gear information required in the RDBES}
\usage{
get_gear_info(
  df = samp,
  year = year,
  area = dfuArea,
  area_code_type = "fishLine",
  gear_type = gearType,
  target_spp = targetSpecies1,
  target_spp_code_type = "fishLine",
  mesh_size = meshSize
)
}
\arguments{
\item{selection_device:}{TO DO}
}
\description{
Get gear information required in the RDBES
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
