% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LO_fishline_2_rdbes.R
\name{getDataModel}
\alias{getDataModel}
\title{FishLine 2 RDBES, Location selection (LO)}
\usage{
getDataModel(tableName)
}
\arguments{
\item{path_to_data_model_baseTypes}{Where to find the baseTypes for the data model}

\item{year}{years needed}

\item{cruises}{Name of cruises in national database}

\item{type}{only_mandatory | everything}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19.13
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
