% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_fishline_2_rdbes_1_17.R
\name{SA_fishline_2_rdbes}
\alias{SA_fishline_2_rdbes}
\title{FishLine 2 RDBES, sample (SA)}
\usage{
SA_fishline_2_rdbes(
  data_model = fields_sequence,
  year = 2019,
  cruises = c("MON", "SEAS")
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{cruises}{Name of cruises in national database}

\item{landingCategory_for_sa}{}
}
\value{

}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.17. This function ...
}
\examples{



}
