% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_get_data.R
\name{SA_get_data}
\alias{SA_get_data}
\title{SA_get_data}
\usage{
SA_get_data(
  link_data = link,
  sampling_scheme = "DNK_AtSea_Observer_Active",
  years = 2021
)
}
\arguments{
\item{xx}{}
}
\description{
SA_get_data
}
