% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_fishline_2_rdbes_1_17.R
\name{SD_fishline_2_rdbes}
\alias{SD_fishline_2_rdbes}
\title{FishLine 2 RDBES, sampling details (SD)}
\usage{
SD_fishline_2_rdbes(data_model = data_model, id_input_file = DE)
}
\arguments{
\item{id_input_file}{}
}
\value{

}
\description{
FishLine 2 RDBES, sampling details (SD)
}
