% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_fishline_2_rdbes.R,
%   R/SL_fishline_2_rdbes_1_17.R
\name{SL_fishline_2_rdbes}
\alias{SL_fishline_2_rdbes}
\title{FishLine 2 RDBES, species list (SL)}
\usage{
SL_fishline_2_rdbes(
  data_model = data_model,
  years = c(2010, 2019),
  year = 2016,
  cruises = c("MON", "SEAS"),
  treshold = 0.05
)

SL_fishline_2_rdbes(
  data_model = data_model,
  years = c(2010, 2019),
  year = 2016,
  cruises = c("MON", "SEAS"),
  treshold = 0.05
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{year}{Only takes a single year for now}

\item{cruises}{Name of cruises in national database}

\item{treshold}{}

\item{type}{only_mandatory | everything}
}
\value{



}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.18. This function ...

Converts sample data from national database (fishLine) to RDBES. Data model v. 1.17. This function ...
}
\examples{






}
