% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_fishline_2_rdbes_1_17.R
\name{SS_fishline_2_rdbes}
\alias{SS_fishline_2_rdbes}
\title{FishLine 2 RDBES, species selection (SS)}
\usage{
SS_fishline_2_rdbes(
  data_model = data_model,
  year = 2016,
  cruises = c("MON", "SEAS"),
  hierarchy = 1,
  input_file = FO[[1]],
  sl_file = SL[[1]]
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{cruises}{Name of cruises in national database}

\item{hierarchy}{}
}
\value{

}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.17. This function ...
}
\examples{



}
