% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VD_fishline_2_rdbes.R
\name{VD_fishline_2_rdbes}
\alias{VD_fishline_2_rdbes}
\title{FishLine 2 RDBES, Vessel details (VD)}
\usage{
VD_fishline_2_rdbes(
  year = 2022,
  cruises = c("MON", "SEAS", "IN-HIRT", "IN-LYNG")
)
}
\arguments{
\item{year}{sample year}

\item{cruises}{string with names of cruises in national database}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
