% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VD_fishline_2_rdbes_1_18.R
\name{VD_fishline_2_rdbes_1_18}
\alias{VD_fishline_2_rdbes_1_18}
\title{FishLine 2 RDBES, biological variables (BV)}
\usage{
VD_fishline_2_rdbes_1_18(
  data_model = fields_sequence,
  year = 2016,
  cruises = c("MON", "SEAS"),
  type = "only_mandatory"
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{year}{Only takes a single year for now}

\item{cruises}{Name of cruises in national database}

\item{type}{only_mandatory | everything}
}
\value{

}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.17. This function ...
}
\examples{



}
