% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS_fishline_2_rdbes.R
\name{VS_fishline_2_rdbes}
\alias{VS_fishline_2_rdbes}
\title{FishLine 2 RDBES, Vessel selection (VS)}
\usage{
VS_fishline_2_rdbes(
  data_model_baseTypes_path = "Q:/mynd/RDB/create_RDBES_data/references",
  encryptedVesselCode_path =
    "Q:/mynd/RDB/create_RDBES_data/RDBES_data_call_2021/output/for_production",
  years = 2016,
  cruises = c("MON", "SEAS", "IN-HIRT"),
  type = "everything"
)
}
\arguments{
\item{cruises}{Name of cruises in national database}

\item{type}{only_mandatory | everything}

\item{path_to_data_model_baseTypes}{Where to find the baseTypes for the data model}

\item{year}{years needed}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
