% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BV_fishline_2_rdbes.R
\name{BV_fishline_2_rdbes}
\alias{BV_fishline_2_rdbes}
\title{FishLine 2 RDBES, Frequency Measure (FM)}
\usage{
BV_fishline_2_rdbes(
  ref_path = "Q:/mynd/RDB/create_RDBES_data/references",
  sampling_scheme = "DNK_Market_Sampling",
  years = 2016,
  bv_measurement = "length",
  type = "everything"
)
}
\arguments{
\item{type}{only_mandatory | everything}

\item{path_to_data_model_baseTypes}{Where to find the baseTypes for the data model}

\item{year}{years needed}

\item{cruises}{Name of cruises in national database}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
