% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FT_fishline_2_rdbes_1_17.R
\name{FT_fishline_2_rdbes}
\alias{FT_fishline_2_rdbes}
\title{FishLine 2 RDBES, fishing trip (FT)}
\usage{
FT_fishline_2_rdbes(
  data_model = data_model,
  year = 2016,
  cruises = c("MON", "SEAS")
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{year}{Only takes a single year for now}

\item{cruises}{Name of cruises in national database}
}
\value{

}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.17. This function ...
}
\examples{



}
