% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_fishline_2_rdbes.R
\name{SA_fishline_2_rdbes}
\alias{SA_fishline_2_rdbes}
\title{FishLine 2 RDBES, Landing event (LE)}
\usage{
SA_fishline_2_rdbes(data = SA_data)
}
\arguments{
\item{data}{}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
