% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_fishline_2_rdbes.R
\name{SL_fishline_2_rdbes}
\alias{SL_fishline_2_rdbes}
\title{FishLine 2 RDBES, species list (SL)}
\usage{
SL_fishline_2_rdbes(
  basis_years = c(2016:2020),
  cruises = c("MON", "SEAS"),
  catch_fractions = c("Dis", "Lan"),
  specieslist_name = "DNK_AtSea_Observer_all_species_same_Dis_Lan",
  species_to_add = c(148776, 137117)
)
}
\arguments{
\item{basis_years}{How many years should the species list be based on}

\item{cruises}{Which cruises are relevant}

\item{catch_fractions}{Catch fractions in the species list}

\item{specieslist_name}{Output name of the species list}

\item{species_to_add}{additinal AphaiaIds not included in the observations}

\item{data_model_path}{Where to find the .xlsx with the data model}
}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.18. This function ...
}
