% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_fishline_2_rdbes.R
\name{SL_fishline_2_rdbes}
\alias{SL_fishline_2_rdbes}
\title{FishLine 2 RDBES, species list (SL)}
\usage{
SL_fishline_2_rdbes(
  ref_path = "Q:/mynd/RDB/create_RDBES_data/references",
  sampling_scheme = "DNK_Market_Sampling",
  years = 2016,
  basis_years = c(2016:2020),
  catch_fractions = c("Dis", "Lan"),
  specieslist_name = "DNK_AtSea_Observer_all_species_same_Dis_Lan",
  species_to_add = c(148776, 137117),
  type = "everything"
)
}
\arguments{
\item{type}{only_mandatory | everything}

\item{data_model}{Where to find the .xlsx with the data model}

\item{year}{Only takes a single year for now}

\item{cruises}{Name of cruises in national database}
}
\description{
Converts samples data from national database (fishLine) to RDBES. Data model v. 1.18. This function ...
}
