% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_fishline_2_rdbes.R
\name{TE_fishline_2_rdbes}
\alias{TE_fishline_2_rdbes}
\title{FishLine 2 RDBES, Species selection (SS)}
\usage{
TE_fishline_2_rdbes(
  ref_path = "Q:/mynd/kibi/RDBES/create_RDBES_data/references",
  sampling_scheme = "DNK_Market_Sampling",
  years = 2016
)
}
\arguments{
\item{path_to_data_model_baseTypes}{Where to find the baseTypes for the data model}

\item{year}{years needed}

\item{cruises}{Name of cruises in national database}

\item{type}{only_mandatory | everything}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
