% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VD_fishline_2_rdbes.R
\name{VD_fishline_2_rdbes}
\alias{VD_fishline_2_rdbes}
\title{FishLine 2 RDBES, Vessel details (VD)}
\usage{
VD_fishline_2_rdbes(
  year = 2024,
  cruises = c("MON", "SEAS", "IN-HIRT", "IN-LYNG"),
  id_type = "without_version"
)
}
\arguments{
\item{year}{sample year}

\item{cruises}{string with names of cruises in national database}

\item{id_type}{with or without version in the encrypted vessel id}
}
\description{
Converts samples data from national database (fishLine) to RDBES.
}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
