% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VD_fishline_2_rdbes.R,
%   R/VD_fishline_2_rdbes_1_17.R
\name{VD_fishline_2_rdbes}
\alias{VD_fishline_2_rdbes}
\title{FishLine 2 RDBES, Vessel details (VD)}
\usage{
VD_fishline_2_rdbes(
  data_model = data_model,
  year = 2016,
  cruises = c("MON", "SEAS"),
  type = "only_mandatory"
)

VD_fishline_2_rdbes(
  data_model = data_model,
  year = 2016,
  cruises = c("MON", "SEAS"),
  type = "only_mandatory"
)
}
\arguments{
\item{data_model}{Where to find the .xlsx with the data model}

\item{year}{Only takes a single year for now}

\item{cruises}{Name of cruises in national database}

\item{type}{only_mandatory | everything}

\item{path_to_data_model_baseTypes}{Where to find the baseTypes for the data model}
}
\value{



}
\description{
Converts samples data from national database (fishLine) to RDBES.
Data model v. 1.19

Converts samples data from national database (fishLine) to RDBES. Data model v. 1.17. This function ...
}
\examples{






}
\author{
Kirsten Birch Håkansson, DTU Aqua
}
